Preliminary support for build.
This puts in place preliminary support for the build command.  This is
based on improvements to JBuildGraph.  However, there remain a number of
outstanding problems.  Firstly, its not clear how a Build.Platform
should be initialised; secondly, its not clear how a Build.Task should
be applied.  More specifically, what contextual information is required.

However, the rough system is in place, there are just some details to be
ironed out in jbuildgraph.