Happy path staring to work!
Downloading of Maven deps is now working on the happy path, though requires
testing on the unhappy paths.  Also not clear that I have all the required
dependencies listed at this stage, but that's fine.

Need to work on error handling, and optimise in a few places.  Also need to
figure out how to manage the global wy.toml file along with the list of plugins.