Support Build Platforms
This adds a notion of a "build" platform which is a trait that can be
instantiated by different concrete platforms (e.g. for whiley, boogie,
etc).  This provides a relatively extensible way of managing different
backends, though falls short of being a true plugin system.  But, it
will do for the time being.

At this stage, there remain some problems.  In particular, the is
currently no mechanism for instantiating a platform for a given
configuration.  Therefore we need a "Descriptor" which accepts something
like an Option<Value> and produces an instance of Platform.