Working on WhileyPath
Currently, there are two things going on:

(1) Adding support for passing whileypath to Wyc

(2) Adding support for passing whileypath to WyJS (and also for it to accept
that, and translate contained WyilFiles and extract an js includes).

At this stage, whiley.rs is now building the WhileyPath as a Vec<String>.  I
think it should be a Vec<PathBuf> really, and there should be a generic way of
extracting it from the config (i.e. since other platforms will need to do the
same thing).  For example, config.get_dependencies() or similar.

To put it altogether I still need to deploy WyC and WyJS as well, and presumably
update versions, etc.