Support folders in Manifest #31
In addition, have updated clean / install / build commands to behave properly.
For example, when building we need to make sure that binary folders exist.
There still remain some issues I think as to what is included in a PACKAGE and
what is not.  For example, currently boogie bpl files are included in the
package which really doesn't make sense.  At the same time, whilst we don't want
generated JavaScript files to be included in a package we do want other
JavaScript files (e.g. native code).