Improved error reporting
This puts in place some improved error reporting.  In particular when plugins
generate output that cannot be parsed, then it will generate an internal
failure.  To be fair, more useful informatio could still be reported.
Nevertheless, it is a good improvement upon silently dropping the error
information!

The next step would be to add better sanity checking to wybt itself.  For
example, when a dependency is specified which does not exist.  This will come
after we add support for the Whiley Package manager.